fetch prices of tokens in all active orders
